Draw logo on startup
